WIP: Trying to drive configure,build,install, and test in test framework...
There appears to be no way to add dependencies from one test to the
next with unit test. So that is a major problem with this approach.

What is the best way to build a infrastructure to easily test this
sequence? Do we need to drive each test module with tox or something?