setuptools_wrap: Add support for pure python packages and packages built by CMake
Prior to this commit, install rules for all packages were required
to be added to the CMake project. This commit changes this by allowing
mixing of pure python package along with package built by CMake.

To illustrate this, the "hello" sample project has been updated to include
a pure python package named 'bonjour' that doesn't have any install
rules in any of the CMakeLists.txt