setuptools_wrap: Add support for "cmake" and "hybrid" python package
An "hybrid" package is a python package that have some files living
in the project source tree and some files installed by CMake.

A "CMake" package is a python package that is fully generated and
installed by CMake without any of his files existing in the source
tree.

The proper functioning of skbuild depends on:
  value of 'cmake_source_dir'
  value of 'cmake_install_dir'
  layout of the source tree

A new test named "test_setup_inputs" has been added to test different
configurations.