ci: Use "scikit-ci.yml" to driver appveyor/circle/travis builds
Commands specific to this project for installing prerequisites, checking
the style, building the project, running the tests, computing the coverage
and creating wheels are now centralized in scikit-ci.yml.

There is no need to maintain three set of commands in each CI settings file.

The commands within appveyor.yml, circle.yml and .travis.yml are now all
similar and simply invoke a driver script that will read "scikit-ci.yml"
and invoke the install/build/test commands specific to this project.