Test generator candidates with call(list), not str
Instead of building a command string for subprocess.call, build the
arguments as a list. That way, subprocess properly handles spaces in
arguments and paths.

The problem with the string-based invocation of subprocess is uses of
skbuild such as

  setup.py build --cmake-executable "C:/Program Files/cmake.exe" ..

fails awkwardly (without actually stopping configuration) with the error
output:

'C:/Program' is not recognized as an internal or external command,
operable program or batch file.

This does not occur when the executable and arguments are passed as a
list.