targetLinkLibrariesWithDynamicLookup: Fix support for cross-compilation exception
This commit fixes dynamic linking when cross-compiling on macOS.

It revisits the approach used to support cross-compilation when
no emulator is available. This is done by further improving on
fixes integrated through c0ec84864 (Fix targetLinkLibrariesWithDynamicLookup)
and 07b2c4ff0 (fix: assume success for osx_dynamic_lookup when cross compiling)

Messages reported during configuration are also improved, when the
compile tests are assumed to pass when cross-compiling without emulator,
a clear message is reported.

For example:

  Performing Test Weak Link MODULE -> SHARED (osx_dynamic_lookup) - Assuming Success (Cross compiling without emulator on macOS)

Co-authored-by: Jean-Christophe Fillion-Robin <jchris.fillionr@kitware.com>