fix: validate value before returning library path (#942)
* fix: validate value before returning library path

In CMaker._guess_python_library(), there is a path where python_library
can be set to a file name without a path. For example, this happens
when using a venv in a MinGW envronment on Windows. This gets passed
to CMake via Python_LIBRARY which causes CMake to fail.

To fix it, we validate that the path exists on disk before returning
the value or return None if it does not exist.

Fixes: https://github.com/scikit-build/scikit-build/issues/940

* Update cmaker.py

* Update cmaker.py

---------

Co-authored-by: Henry Schreiner <HenrySchreinerIII@gmail.com>