fix: Update README.rst with known issue on AIX
AIX support requires a version of CMake newer than the 3.22.0 version that
IBM supplies in their AIX Toolbox for Open Source Software.  IBM is aware
of the issue and will be providing a newer build soon, but for now we
should note that users will need to build their own copy of a newer CMake
on AIX at this time.